+++ src/platforms/wordpress/classes/Gantry/WordPress/Widget/Particle.php
@@ -70,8 +70,11 @@ class Particle extends \WP_Widget
         if (!is_array($instance)) {
             $instance = [];
         }
+
+        $widget_id = isset($args['widget_id']) ? preg_replace('/[^\d]/', '', $args['widget_id']) : null;
         $md5 = md5(json_encode($instance));
-        $id = isset($instance['id']) ? $instance['id'] : "widget-{$md5}";
+        $id = isset($instance['id']) ? $instance['id'] : ($widget_id ?: "widget-{$md5}");
+
         if (!isset($this->content[$md5])) {
             /** @var Theme $theme */
             $theme = $this->container['theme'];
@@ -109,6 +112,9 @@ class Particle extends \WP_Widget
                 'gantry' => $this->container,
                 'inContent' => true
             );
+            if (isset($args['ajax'])) {
+                $context['ajax'] = $args['ajax'];
+            }
 
             $this->content[$md5] = $theme->getContent($object, $context);
         }