+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # 5.4.15
 ## mm/dd/2017
 
-1. [Common](#common)
+2. [Joomla](#joomla)
+    1. [](#new)
+        - Add AJAX support for particles (#1376)
 
 # 5.4.14
 ## 07/06/2017