+++ CHANGELOG.md
@@ -5,12 +5,18 @@
     1. [](#new)
         - Added AJAX support for particles (#1376)
         - Added new twig filter `|attribute_array` to convert array of key => attribute into HTML
+        - Helium/Hydrogen: added missing styling for disabled button state
     2. [](#improved)
         - Improved usability by changing order of Tabs in admin (#2061)
     3. [](#bugfix)
         - Fixed bug in nested `collection.list` on `container.tabs` (#1995)
         - Fixed no space between block attributes (#2074)
-        
+2. [Joomla](#joomla)
+    2. [](#improved)
+        - Joomla Articles particle won't display images alt tags (#2076)
+    3. [](#bugfix)
+        - Removed forced input heights set on 19px (#2063)
+
 # 5.4.14
 ## 07/06/2017